+++ ui/apps/platform/src/Containers/Vulnerabilities/hooks/useInvalidateVulnerabilityQueries.ts
@@ -2,7 +2,7 @@ import { useApolloClient } from '@apollo/client';
 import { DocumentNode, FieldNode, Kind, OperationDefinitionNode, OperationTypeNode } from 'graphql';
 import uniq from 'lodash/uniq';
 
-import { imageListQuery } from '../WorkloadCves/Tables/ImagesTable';
+import { imageListQuery } from '../WorkloadCves/Tables/ImageOverviewTable';
 import { cveListQuery } from '../WorkloadCves/Tables/WorkloadCVEOverviewTable';
 import { deploymentListQuery } from '../WorkloadCves/Tables/DeploymentOverviewTable';
 import { deploymentVulnerabilitiesQuery } from '../WorkloadCves/Deployment/DeploymentPageVulnerabilities';
